+++ Cineaste/ViewController/MovieNight/MovieNightViewController.swift
@@ -244,7 +244,7 @@ class MovieNightViewController: UITableViewController {
     }
 
     private func generateOwnNearbyMessage() -> NearbyMessage {
-        guard let username = UserDefaultsManager.getUsername()
+        guard let username = UsernamePersistence.username
             else { fatalError("ViewController should never be presented without a username") }
 
         let nearbyMovies = movies