+++ Cineaste/ViewController/Settings/SettingsViewController.swift
@@ -47,6 +47,14 @@ class SettingsViewController: UITableViewController {
         store.unsubscribe(self)
     }
 
+    override func viewDidAppear(_ animated: Bool) {
+        super.viewDidAppear(animated)
+
+        guard let rowForUsername = settings.firstIndex(of: SettingItem.name) else { return }
+        let indexPath = IndexPath(row: rowForUsername, section: 0)
+        tableView.reloadRows(at: [indexPath], with: .none)
+    }
+
     override func viewDidLayoutSubviews() {
         super.viewDidLayoutSubviews()
 
@@ -101,6 +109,19 @@ extension SettingsViewController {
         docController = UIDocumentInteractionController(url: url)
         docController?.uti = String.exportMoviesFileUTI
         docController?.presentOptionsMenu(from: rect, in: view, animated: true)
+    }
+}
+
+// MARK: - UITextFieldDelegate
+
+extension SettingsViewController: UITextFieldDelegate {
+    func textField(_ textField: UITextField, shouldChangeCharactersIn range: NSRange, replacementString string: String) -> Bool {
+        guard let text = textField.text else { return true }
+
+        let entryLength = text.count + string.count - range.length
+        UsernameAlert.saveAction?.isEnabled = entryLength > 0
+
+        return true
     }
 }